修改：     include/caffe2/util/model.h 	修改：     include/caffe2/util/net.h 	修改：     src/caffe2/util/model.cc 	修改：     src/caffe2/util/net_operator.cc

+++ include/caffe2/util/net.h
@@ -139,7 +139,7 @@ class NetUtil {
                         const std::string& output);
   OperatorDef* AddPowOp(const std::string & input, const std::string & output, float exponent);
   OperatorDef* AddSubOp(const std::vector<std::string> & inputs, const std::string & output, int broadcast = 1, int axis = 1);
-  OperatorDef* AddUpsampleNearestOp(const std::string & input,const std::string & output, float scale = 2);
+  OperatorDef* AddUpsampleNearestOp(const std::string & input,const std::string & output, int scale = 2);
 
   OperatorDef* AddLSTMUnitOp(const std::vector<std::string>& inputs,
                              const std::vector<std::string>& outputs,
